watchdog: Update host watchdog event name triggered
When the watchdog timer hits, if the hostboot-dump-collection is
disabled, the org.open_power.Host.Boot.Error.WatchdogTimedOut event is
triggered.

If the hostboot-dump-collection is enabled, the
org.open_power.Host.Boot.Error.WatchdogTimeout event is triggered

Due to this, when host dump collection is disabled, the event did not
match any entry in registry entry. Changing the event name and the
message registry in both enabled and disabled cases to
org.open_power.Host.Boot.Error.WatchdogTimedOut to see the expected PEL.

Fix: Update the event name to
org.open_power.Host.Boot.Error.WatchdogTimedOut