opdreport: Eliminate undeclared variable; use existing variable instead
The variable 'id' is not declared anywhere in the code, but it is
used for determining the dump content type in the code. Because of
this, dump content type is not populated correctly which is causing
a failure in determining dump type when the dump is parsed using
'makedump' tool. So, removing usage of variable 'id' and using
'dump_id' instead of it to solve this issue.

Test Results:

Before:
>>>> Processing dump files
  Processing: SYSDUMP.13E09F0.00000007.20240913135832
  Dump Type is: UNKNOWN

>>>> Processing dump files
  Processing: SYSDUMP.13ECF8X.20000008.20240916083507
  Dump Type is: UNKNOWN   

After:
>>>> Processing dump files
  Processing: SYSDUMP.1012345.00000008.20240903113715
  Dump Type is: SCS

>>>> Processing dump files
  Processing: SYSDUMP.1012345.20000007.20240903113517
  Dump Type is: HB

>>>> Processing dump files
  Processing: SYSDUMP.1012345.20000010.20240903140425
  Dump Type is: HB

>>>> Processing dump files
  Processing: SYSDUMP.1012345.00000011.20240903140502
  Dump Type is: SCS

Change-Id: Ic0a85b2a0dee43629dde920c5f13af05e6fb8347
Signed-off-by: SwethaParasa <parasa.swetha1@ibm.com>